requirements.txt
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
multidict==6.0.4
pycares==4.4.0
pycparser==2.21
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2023.3.post1
sniffio==1.3.0
yarl==1.9.4